Parameterize pseudo random bytes
Remove helpers module.
Move necessary functions from old helpers module to hash_to_field
module.
Remove sha3 dependency.
