+++ crypto/src/hash/mod.rs
@@ -1,5 +1,4 @@
 pub mod hash_to_field;
-pub mod helpers;
 pub mod poseidon;
 pub mod sha3;
 pub mod traits;
